Add OAuth over PLAIN tests to keycloak-auth-test + refactor keycloak-auth-test
