--- testsuite/keycloak-auth-test/src/test/java/io/strimzi/testsuite/oauth/KeycloakAuthenticationTest.java
@@ -1,345 +1,0 @@
-/*
- * Copyright 2017-2020, Strimzi authors.
- * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
- */
-package io.strimzi.testsuite.oauth;
-
-import com.fasterxml.jackson.databind.JsonNode;
-import io.strimzi.kafka.oauth.client.ClientConfig;
-import io.strimzi.kafka.oauth.common.HttpUtil;
-import io.strimzi.kafka.oauth.common.OAuthAuthenticator;
-import io.strimzi.kafka.oauth.common.TokenInfo;
-import org.apache.kafka.clients.consumer.Consumer;
-import org.apache.kafka.clients.consumer.ConsumerConfig;
-import org.apache.kafka.clients.consumer.ConsumerRecords;
-import org.apache.kafka.clients.consumer.KafkaConsumer;
-import org.apache.kafka.clients.producer.KafkaProducer;
-import org.apache.kafka.clients.producer.Producer;
-import org.apache.kafka.clients.producer.ProducerConfig;
-import org.apache.kafka.clients.producer.ProducerRecord;
-import org.apache.kafka.clients.producer.RecordMetadata;
-import org.apache.kafka.common.TopicPartition;
-import org.apache.kafka.common.errors.AuthenticationException;
-import org.apache.kafka.common.serialization.StringDeserializer;
-import org.apache.kafka.common.serialization.StringSerializer;
-import org.jboss.arquillian.junit.Arquillian;
-import org.junit.Assert;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import java.io.IOException;
-import java.net.URI;
-import java.time.Duration;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Properties;
-import java.util.concurrent.ExecutionException;
-
-/**
- * Tests for OAuth authentication using Keycloak
- *
- * This test assumes there are multiple listeners configured with OAUTHBEARER support, but each configured differently
- * - configured with different options, or different realm.
- *
- * There should be no authorization configured on the Kafka broker.
- */
-@RunWith(Arquillian.class)
-public class KeycloakAuthenticationTest {
-
-    @Test
-    public void doTest() throws Exception {
-        clientCredentialsWithJwtECDSAValidation();
-        accessTokenWithIntrospectionTest();
-        refreshTokenWithIntrospectionTest();
-        clientCredentialsWithJwtAudienceTest();
-        clientCredentialsWithIntrospectionAudienceTest();
-    }
-
-    public void clientCredentialsWithJwtECDSAValidation() throws Exception {
-
-        System.out.println("==== KeycloakAuthenticationTest :: clientCredentialsWithJwtECDSAValidationTest ====");
-
-        final String kafkaBootstrap = "kafka:9092";
-        final String hostPort = "keycloak:8080";
-        final String realm = "demo-ec";
-
-        final String tokenEndpointUri = "http://" + hostPort + "/auth/realms/" + realm + "/protocol/openid-connect/token";
-
-        Map<String, String> oauthConfig = new HashMap<>();
-        oauthConfig.put(ClientConfig.OAUTH_TOKEN_ENDPOINT_URI, tokenEndpointUri);
-        oauthConfig.put(ClientConfig.OAUTH_CLIENT_ID, "kafka-producer-client");
-        oauthConfig.put(ClientConfig.OAUTH_CLIENT_SECRET, "kafka-producer-client-secret");
-        oauthConfig.put(ClientConfig.OAUTH_USERNAME_CLAIM, "preferred_username");
-
-        Properties producerProps = buildProducerConfigOAuthBearer(kafkaBootstrap, oauthConfig);
-        Producer<String, String> producer = new KafkaProducer<>(producerProps);
-
-        final String topic = "KeycloakAuthenticationTest-clientCredentialsWithJwtECDSAValidationTest";
-
-
-        producer.send(new ProducerRecord<>(topic, "The Message")).get();
-        System.out.println("Produced The Message");
-
-        Properties consumerProps = buildConsumerConfigOAuthBearer(kafkaBootstrap, oauthConfig);
-        Consumer<String, String> consumer = new KafkaConsumer<>(consumerProps);
-
-        TopicPartition partition = new TopicPartition(topic, 0);
-        consumer.assign(Collections.singletonList(partition));
-
-        while (consumer.partitionsFor(topic, Duration.ofSeconds(1)).size() == 0) {
-            System.out.println("No assignment yet for consumer");
-        }
-        consumer.seekToBeginning(Collections.singletonList(partition));
-
-        ConsumerRecords<String, String> records = consumer.poll(Duration.ofSeconds(1));
-
-        Assert.assertEquals("Got message", 1, records.count());
-        Assert.assertEquals("Is message text: 'The Message'", "The Message", records.iterator().next().value());
-    }
-
-    public void accessTokenWithIntrospectionTest() throws Exception {
-        System.out.println("==== KeycloakAuthenticationTest :: accessTokenWithIntrospectionTest ====");
-
-        final String kafkaBootstrap = "kafka:9093";
-        final String hostPort = "keycloak:8080";
-        final String realm = "demo";
-
-        final String tokenEndpointUri = "http://" + hostPort + "/auth/realms/" + realm + "/protocol/openid-connect/token";
-
-        final String clientId = "kafka-producer-client";
-        final String clientSecret = "kafka-producer-client-secret";
-
-        // first, request access token using client id and secret
-        TokenInfo info = OAuthAuthenticator.loginWithClientSecret(URI.create(tokenEndpointUri), null, null, clientId, clientSecret, true, null, null);
-
-        Map<String, String> oauthConfig = new HashMap<>();
-        oauthConfig.put(ClientConfig.OAUTH_TOKEN_ENDPOINT_URI, tokenEndpointUri);
-        oauthConfig.put(ClientConfig.OAUTH_ACCESS_TOKEN, info.token());
-        oauthConfig.put(ClientConfig.OAUTH_USERNAME_CLAIM, "preferred_username");
-
-        Properties producerProps = buildProducerConfigOAuthBearer(kafkaBootstrap, oauthConfig);
-        Producer<String, String> producer = new KafkaProducer<>(producerProps);
-
-
-        final String topic = "KeycloakAuthenticationTest-accessTokenWithIntrospectionTest";
-
-        producer.send(new ProducerRecord<>(topic, "The Message")).get();
-        System.out.println("Produced The Message");
-
-        Properties consumerProps = buildConsumerConfigOAuthBearer(kafkaBootstrap, oauthConfig);
-        Consumer<String, String> consumer = new KafkaConsumer<>(consumerProps);
-
-        TopicPartition partition = new TopicPartition(topic, 0);
-        consumer.assign(Collections.singletonList(partition));
-
-        while (consumer.partitionsFor(topic, Duration.ofSeconds(1)).size() == 0) {
-            System.out.println("No assignment yet for consumer");
-        }
-        consumer.seekToBeginning(Collections.singletonList(partition));
-
-        ConsumerRecords<String, String> records = consumer.poll(Duration.ofSeconds(1));
-
-        Assert.assertEquals("Got message", 1, records.count());
-        Assert.assertEquals("Is message text: 'The Message'", "The Message", records.iterator().next().value());
-    }
-
-    public void refreshTokenWithIntrospectionTest() throws Exception {
-
-        System.out.println("==== KeycloakAuthenticationTest :: refreshTokenWithIntrospectionTest ====");
-
-        final String kafkaBootstrap = "kafka:9093";
-        final String hostPort = "keycloak:8080";
-        final String realm = "demo";
-
-        final String tokenEndpointUri = "http://" + hostPort + "/auth/realms/" + realm + "/protocol/openid-connect/token";
-
-        final String clientId = "kafka-cli";
-        final String username = "alice";
-        final String password = "alice-password";
-
-        // first, request access token using client id and secret
-        String refreshToken = loginWithUsernameForRefreshToken(URI.create(tokenEndpointUri), username, password, clientId);
-
-        Map<String, String> oauthConfig = new HashMap<>();
-        oauthConfig.put(ClientConfig.OAUTH_TOKEN_ENDPOINT_URI, tokenEndpointUri);
-        oauthConfig.put(ClientConfig.OAUTH_CLIENT_ID, clientId);
-        oauthConfig.put(ClientConfig.OAUTH_REFRESH_TOKEN, refreshToken);
-        oauthConfig.put(ClientConfig.OAUTH_USERNAME_CLAIM, "preferred_username");
-
-        Properties producerProps = buildProducerConfigOAuthBearer(kafkaBootstrap, oauthConfig);
-        Producer<String, String> producer = new KafkaProducer<>(producerProps);
-
-
-        final String topic = "KeycloakAuthenticationTest-refreshTokenWithIntrospectionTest";
-
-        producer.send(new ProducerRecord<>(topic, "The Message")).get();
-        System.out.println("Produced The Message");
-
-        Properties consumerProps = buildConsumerConfigOAuthBearer(kafkaBootstrap, oauthConfig);
-        Consumer<String, String> consumer = new KafkaConsumer<>(consumerProps);
-
-        TopicPartition partition = new TopicPartition(topic, 0);
-        consumer.assign(Collections.singletonList(partition));
-
-        while (consumer.partitionsFor(topic, Duration.ofSeconds(1)).size() == 0) {
-            System.out.println("No assignment yet for consumer");
-        }
-        consumer.seekToBeginning(Collections.singletonList(partition));
-
-        ConsumerRecords<String, String> records = consumer.poll(Duration.ofSeconds(1));
-
-        Assert.assertEquals("Got message", 1, records.count());
-        Assert.assertEquals("Is message text: 'The Message'", "The Message", records.iterator().next().value());
-    }
-
-    public void clientCredentialsWithJwtAudienceTest() throws Exception {
-        System.out.println("==== KeycloakAuthenticationTest :: clientCredentialsWithJwtAudienceTest ====");
-
-        final String kafkaBootstrap = "kafka:9094";
-        final String hostPort = "keycloak:8080";
-        final String realm = "kafka-authz";
-
-        final String tokenEndpointUri = "http://" + hostPort + "/auth/realms/" + realm + "/protocol/openid-connect/token";
-
-        Map<String, String> oauthConfig = new HashMap<>();
-        oauthConfig.put(ClientConfig.OAUTH_TOKEN_ENDPOINT_URI, tokenEndpointUri);
-        oauthConfig.put(ClientConfig.OAUTH_CLIENT_ID, "team-b-client");
-        oauthConfig.put(ClientConfig.OAUTH_CLIENT_SECRET, "team-b-client-secret");
-        oauthConfig.put(ClientConfig.OAUTH_USERNAME_CLAIM, "preferred_username");
-
-        Properties producerProps = buildProducerConfigOAuthBearer(kafkaBootstrap, oauthConfig);
-        Producer<String, String> producer = new KafkaProducer<>(producerProps);
-
-        String topic = "KeycloakAuthenticationTest-clientCredentialsWithJwtAudienceTest";
-
-        RecordMetadata result = producer.send(new ProducerRecord<>(topic, "message")).get();
-
-        Assert.assertTrue("Has offset", result.hasOffset());
-        producer.close();
-
-        System.out.println("Produced The Message");
-
-
-        oauthConfig.put(ClientConfig.OAUTH_CLIENT_ID, "team-a-client");
-        oauthConfig.put(ClientConfig.OAUTH_CLIENT_SECRET, "team-a-client-secret");
-
-        producerProps = buildProducerConfigOAuthBearer(kafkaBootstrap, oauthConfig);
-        producer = new KafkaProducer<>(producerProps);
-        try {
-            producer.send(new ProducerRecord<>(topic, "message2")).get();
-            Assert.fail();
-
-        } catch (ExecutionException e) {
-            Throwable cause = e.getCause();
-            Assert.assertTrue("instanceOf AuthenticationException", cause instanceof AuthenticationException);
-            Assert.assertTrue("audience check failed", cause.toString().contains("audience"));
-        }
-    }
-
-    public void clientCredentialsWithIntrospectionAudienceTest() throws Exception {
-        System.out.println("==== KeycloakAuthenticationTest :: clientCredentialsWithIntrospectionAudienceTest ====");
-
-        final String kafkaBootstrap = "kafka:9095";
-        final String hostPort = "keycloak:8080";
-        final String realm = "kafka-authz";
-
-        final String tokenEndpointUri = "http://" + hostPort + "/auth/realms/" + realm + "/protocol/openid-connect/token";
-
-        Map<String, String> oauthConfig = new HashMap<>();
-        oauthConfig.put(ClientConfig.OAUTH_TOKEN_ENDPOINT_URI, tokenEndpointUri);
-        oauthConfig.put(ClientConfig.OAUTH_CLIENT_ID, "team-b-client");
-        oauthConfig.put(ClientConfig.OAUTH_CLIENT_SECRET, "team-b-client-secret");
-        oauthConfig.put(ClientConfig.OAUTH_USERNAME_CLAIM, "preferred_username");
-
-        Properties producerProps = buildProducerConfigOAuthBearer(kafkaBootstrap, oauthConfig);
-        Producer<String, String> producer = new KafkaProducer<>(producerProps);
-
-        String topic = "KeycloakAuthenticationTest-clientCredentialsWithIntrospectionAudienceTest";
-
-        RecordMetadata result = producer.send(new ProducerRecord<>(topic, "message")).get();
-
-        Assert.assertTrue("Has offset", result.hasOffset());
-        producer.close();
-
-        System.out.println("Produced The Message");
-
-
-        oauthConfig.put(ClientConfig.OAUTH_CLIENT_ID, "team-a-client");
-        oauthConfig.put(ClientConfig.OAUTH_CLIENT_SECRET, "team-a-client-secret");
-
-        producerProps = buildProducerConfigOAuthBearer(kafkaBootstrap, oauthConfig);
-        producer = new KafkaProducer<>(producerProps);
-        try {
-            producer.send(new ProducerRecord<>(topic, "message2")).get();
-            Assert.fail();
-
-        } catch (ExecutionException e) {
-            Throwable cause = e.getCause();
-            Assert.assertTrue("instanceOf AuthenticationException", cause instanceof AuthenticationException);
-            Assert.assertTrue("audience check failed", cause.toString().contains("audience"));
-        }
-
-    }
-
-
-
-
-    private static String getJaasConfigOptionsString(Map<String, String> options) {
-        StringBuilder sb = new StringBuilder();
-        for (Map.Entry<String, String> ent: options.entrySet()) {
-            sb.append(" ").append(ent.getKey()).append("=\"").append(ent.getValue()).append("\"");
-        }
-        return sb.toString();
-    }
-
-    private static Properties buildProducerConfigOAuthBearer(String kafkaBootstrap, Map<String, String> oauthConfig) {
-        Properties p = buildCommonConfigOAuthBearer(kafkaBootstrap, oauthConfig);
-        p.setProperty(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, kafkaBootstrap);
-        p.setProperty(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class.getName());
-        p.setProperty(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, StringSerializer.class.getName());
-        p.setProperty(ProducerConfig.ACKS_CONFIG, "all");
-        return p;
-    }
-
-    private static Properties buildConsumerConfigOAuthBearer(String kafkaBootstrap, Map<String, String> oauthConfig) {
-        Properties p = buildCommonConfigOAuthBearer(kafkaBootstrap, oauthConfig);
-        p.setProperty(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class.getName());
-        p.setProperty(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class.getName());
-        p.setProperty(ConsumerConfig.GROUP_ID_CONFIG, "consumer-group");
-        p.setProperty(ConsumerConfig.MAX_POLL_RECORDS_CONFIG, "10");
-        p.setProperty(ConsumerConfig.ENABLE_AUTO_COMMIT_CONFIG, "true");
-        return p;
-    }
-
-    private static Properties buildCommonConfigOAuthBearer(String kafkaBootstrap, Map<String, String> oauthConfig) {
-        String configOptions = getJaasConfigOptionsString(oauthConfig);
-
-        Properties p = new Properties();
-        p.setProperty("security.protocol", "SASL_PLAINTEXT");
-        p.setProperty("sasl.mechanism", "OAUTHBEARER");
-        p.setProperty("sasl.jaas.config", "org.apache.kafka.common.security.oauthbearer.OAuthBearerLoginModule required " + configOptions + " ;");
-        p.setProperty("sasl.login.callback.handler.class", "io.strimzi.kafka.oauth.client.JaasClientOauthLoginCallbackHandler");
-
-        p.setProperty(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, kafkaBootstrap);
-        return p;
-    }
-
-    private String loginWithUsernameForRefreshToken(URI tokenEndpointUri, String username, String password, String clientId) throws IOException {
-
-        JsonNode result = HttpUtil.post(tokenEndpointUri,
-                null,
-                null,
-                null,
-                "application/x-www-form-urlencoded",
-                "grant_type=password&username=" + username + "&password=" + password + "&client_id=" + clientId,
-                JsonNode.class);
-
-        JsonNode token = result.get("refresh_token");
-        if (token == null) {
-            throw new IllegalStateException("Invalid response from authorization server: no refresh_token");
-        }
-        return token.asText();
-    }
-
-}
