+++ testsuite/keycloak-auth-tests/pom.xml
@@ -10,7 +10,7 @@
         <version>1.0.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>keycloak-auth-test</artifactId>
+    <artifactId>keycloak-auth-tests</artifactId>
 
     <licenses>
         <license>
