+++ testsuite/keycloak-auth-tests/src/test/java/io/strimzi/testsuite/oauth/KeycloakAuthenticationTest.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2017-2020, Strimzi authors.
+ * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
+ */
+package io.strimzi.testsuite.oauth;
+
+import org.jboss.arquillian.junit.Arquillian;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+/**
+ * Tests for OAuth authentication using Keycloak
+ *
+ * This test assumes there are multiple listeners configured with OAUTHBEARER support, but each configured differently
+ * - configured with different options, or different realm.
+ *
+ * There should be no authorization configured on the Kafka broker.
+ */
+@RunWith(Arquillian.class)
+public class KeycloakAuthenticationTest {
+
+    @Test
+    public void doTest() throws Exception {
+        BasicTests.doTests();
+        OAuthOverPlainTests.doTests();
+        AudienceTests.doTests();
+    }
+}
